// src/modules/principal/models/Technology.ts
/**
 * Technology interface represents a technology stack or tool
 * Used to display technology information in portfolio sections
 */
export interface Technology {
  /**
   * The display name of the technology
   */
  name: string;

  /**
   * The icon representation of the technology
   */
  icon: ImageMetadata;
}